random names pick until number of time prompted is reached
Randomly choose a name from a list and covert the names on list with the value of times it has been picked. Get input from a user of how many times that name should be picked. 'When the value of that number is reached in the dictionary values the program will end.