random choice of name values put to dictionary
Ask user to give a number  1-10 input(prompt)
Randomly choose name from a list. Create dictionary with name of list with value of zero.
Each time the name is picked the value is stored in the dictionary. 
When the value is reach of the given input; the program is ended